refactor(): type
