+++ src/intent/index.ts
@@ -1,8 +1,23 @@
-import type {CardTokenRequest, StripeAccountIdOpt} from '../definitions';
+import type {CardTokenRequest} from '../definitions';
+import type {StripeAccountIdOpt} from '../shared';
+
+import type { ConfirmPaymentIntentOptions } from './confirm.interface'
+import type { ConfirmPaymentIntentResponse } from './payment.inteface'
+import type { ConfirmSetupIntentOptions, ConfirmSetupIntentResponse } from './setup.interface'
 
 export * from './confirm.interface';
 export * from './setup.interface';
 export * from './payment.inteface'
+
+export interface IntentDefinitions {
+  confirmPaymentIntent(
+    opts: ConfirmPaymentIntentOptions,
+  ): Promise<ConfirmPaymentIntentResponse>;
+
+  confirmSetupIntent(
+    opts: ConfirmSetupIntentOptions,
+  ): Promise<ConfirmSetupIntentResponse>;
+}
 
 export interface CommonIntentOptions extends StripeAccountIdOpt {
   clientSecret: string;
